+++ src/MicroElements.Metadata/Metadata/IPropertyValueFactory.cs
@@ -1,6 +1,10 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+
 namespace MicroElements.Metadata
 {
     /// <summary>
@@ -11,10 +15,43 @@ namespace MicroElements.Metadata
         /// <summary>
         /// Creates new instance if <see cref="IPropertyValue"/>.
         /// </summary>
+        /// <typeparam name="T">Value type.</typeparam>
         /// <param name="property">Property.</param>
         /// <param name="value">Value for property.</param>
         /// <param name="valueSource">Value source.</param>
         /// <returns>Created property value.</returns>
-        IPropertyValue Create(IProperty property, object? value, ValueSource? valueSource = null);
+        IPropertyValue<T> Create<T>(IProperty<T> property, T? value, ValueSource? valueSource = null);
+
+        /// <summary>
+        /// Creates new instance if <see cref="IPropertyValue"/>.
+        /// </summary>
+        /// <param name="property">Property.</param>
+        /// <param name="value">Value for property.</param>
+        /// <param name="valueSource">Value source.</param>
+        /// <returns>Created property value.</returns>
+        IPropertyValue CreateUntyped(IProperty property, object? value, ValueSource? valueSource = null);
+    }
+
+    public interface IPropertyValueFactoryProvider
+    {
+        IPropertyValueFactory GetFactory(IEqualityComparer<IProperty> propertyComparer);
+    }
+
+    public class PropertyValueFactoryProvider : IPropertyValueFactoryProvider
+    {
+        ConcurrentDictionary<IEqualityComparer<IProperty>, IPropertyValueFactory> _factories = new ConcurrentDictionary<IEqualityComparer<IProperty>, IPropertyValueFactory>();
+
+        private Func<IEqualityComparer<IProperty>, IPropertyValueFactory> _factoryFactory;
+
+        public PropertyValueFactoryProvider(Func<IEqualityComparer<IProperty>, IPropertyValueFactory> factoryFactory)
+        {
+            _factoryFactory = factoryFactory;
+        }
+
+        /// <inheritdoc />
+        public IPropertyValueFactory GetFactory(IEqualityComparer<IProperty> propertyComparer)
+        {
+            return _factories.GetOrAdd(propertyComparer, comparer => _factoryFactory(comparer));
+        }
     }
 }