+++ src/MicroElements.Metadata/Metadata/Property.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
-using System.Runtime.CompilerServices;
 using MicroElements.Functional;
 
 namespace MicroElements.Metadata